Reorganize the developer guide and add the section for pull requests
Motivation:

Our developer guide is not organized in the order where a potential
contributor can follow sequentially.  It will be more accessible for a
newcomer if he or she can follow the steps easily.

Modification:

Reorganize the developer guide to repect the usual contribution
workflow.  Also added the guide for issuing a pull request and the
checklist before a push.

Result:

Maybe higher quality pull request?